Refactor: move flow triggering flow code into utils
